[Reporting] Improve errors from failed resolutions in Upgrade Assistant
This change gives more logging and user-friendly error message in the event that the current user does not have the privilege to manage the reporting indices. Previously, there would be a mysterious 404 message. After this change, the server response is a 403 with a more meaningful error message in the response body.