[CI] Enable failed test collection on ES-intake triggered serverless verification runs (#191937)
## Summary
Currently, failed runs on
https://buildkite.com/elastic/kibana-elasticsearch-serverless-verify-and-promote
that were triggered by the ES Serverless pipeline's downstream, are not
collecting failures. This makes it harder to grasp what set of tests
failed.

I think we originally disabled this reporting not to muddy our stats
based on runs on various ES Serverless versions - however, with this,
it's really hard to get an overview on which tests are failing for a
run, so I'd like to get the failed suites to be displayed on the job.

With this change - if I'm not mistaken - we'll generate the annotation
on the Buildkite job, but we won't report to ci-stats or create github
issues.

### In short
What we have:
<img width="889" alt="Screenshot 2024-09-02 at 18 28 45"
src="https://github.com/user-attachments/assets/d294dc00-4dd0-4b29-9e2f-866d68246a5d">

What I'd like:
<img width="1195" alt="Screenshot 2024-09-02 at 18 29 14"
src="https://github.com/user-attachments/assets/455465ba-f176-4ab5-ba31-9578c167448f">

Tested in:
https://buildkite.com/elastic/kibana-elasticsearch-serverless-verify-and-promote/builds/1962

(cherry picked from commit 243de82cc6b7bb7ac7639c65c19147509674b487)