[Synthetics] [ON Week] Refactor Synthetics Overview page for increased scalability (#187092)
## Summary

Resolves https://github.com/elastic/kibana/issues/183622.

This is part of some ON Week PoC work I did that introduces [React
Window](https://github.com/bvaughn/react-window) and its [Infinite
Loader](https://github.com/bvaughn/react-window-infinite-loader) to the
Synthetics Overview page.

Additionally, I have refactored the way that duration fetching is done
for the overview cards; the fetching is now handled at a top-level and
as-needed during the infinite scroll. This results in a page that is
scrollable all the way to the end of very large amounts of monitors. The
example deployment I'm using has about 900 monitors, and in the latest
versions of Kibana this dataset causes the overview page to crash the
tab.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
Co-authored-by: Shahzad <shahzad31comp@gmail.com>