[Observability Onboarding] Add new telemetry hook for auto-detect flow (#191028)
## Summary

Resolves https://github.com/elastic/observability-dev/issues/3856.

Adds a new field to the onboarding telemetry event type and ships two
new events, one for awaiting data and the other for receiving data. The
new field is a string array that will indicate the integration package
names the user has chosen.

~NOTE: I was unsure how much detail to include about the integrations in
question. Right now, the event would ship `integration.title`, but
perhaps we want to re-define the new field to include all the
integration fields and have it contain that?~

There's information in the [comment
below](https://github.com/elastic/kibana/pull/191028#issuecomment-2310613450)
showing the types of data we will collect with these changes.