[Logs] Update logs stream links to logs explorer (#190835)
## Summary

Implements https://github.com/elastic/kibana/issues/190782, as part of
the deprecate Logs UI project:
https://github.com/elastic/observability-dev/issues/3242.

## Overview

- Previously the top level logs locators would redirect to the locators
registered in Infra first (for the stream), and logs explorer second if
the stream wasn't available. This fork no longer exists, with the top
level locators navigating to the logs explorer only.

- Infra no longer registers locators for the logs stream.

- Consumer / solution uses have been updated.

## Semi-breaking change

- Previously we had `link-to` routes that would translate into an
internal link. These no longer accept a log view (as this is a stream
concept, and is also soon to be deprecated itself). This means
bookmarked link-to routes may now ultimately point to slightly different
data (depending on the users log sources advanced setting). I don't
really see a full proof way around this if we are to truly deprecate all
the pieces we need. Also, with the `link-to` routes always being a
translation layer this might be absolutely fine — we are just
translating things differently now.

## Reviewer notes

- The ES Deprecation logs page can be found at:
`/app/management/stack/upgrade_assistant/es_deprecation_logs`.

## Change points

(Not fully exhaustive)

![Screenshot 2024-08-30 at 17 04
24](https://github.com/user-attachments/assets/9062de51-aba1-47bd-b95d-59f72b2202bf)

![Screenshot 2024-08-30 at 17 15
37](https://github.com/user-attachments/assets/c54c91ca-6750-408d-b3cb-d3895a7ae9a1)

![Screenshot 2024-08-30 at 17 18
51](https://github.com/user-attachments/assets/a4e3241d-85ed-4cdc-bb22-4ba07bd077f2)

![Screenshot 2024-08-30 at 17 21
32](https://github.com/user-attachments/assets/2263c463-133f-41d0-ba3e-3943ea48a950)

![Screenshot 2024-08-30 at 17 23
35](https://github.com/user-attachments/assets/0f600a38-5102-40d3-ae1e-4106ee287652)

---------

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>