[react@18] fix `field.helpText` type issue (#192083)
## Summary

Part of https://github.com/elastic/kibana/issues/138222

This is one of the issues that `@types/react@18` upgrade highlights and
that needs to be addressed with or before the upgrade.

`field.helpText` is `ReactNode`, a function is not a valid `ReactNode`,
but `@types/react@17` allowed a function. That is why `typeof
field.helpText === 'function' ? field.helpText() : field.helpText`
doesn't fail with `@17`, but fails with `@18` as `field.helpText is not
callable, as never is not callable`. You can check the types with
@types/react@18 yourself here
https://github.com/elastic/kibana/pull/191738

It looks like the lazy `helpText` isn't needed apart from a hack for
index management where documentation service is used as part of helpText
that could be not available yet.

To keep it supported, I specifically, allowed a function that returns a
ReactNode for `helpText` on `FieldConfig`, but to make consumption clean
and to now break other places `useField` will call it and pass just a
`ReactNode` down