Add new function for importing rules, both prebuilt and custom
Now that we calculate proper values for `rule_source` and `immutable` in
the outer loop of `importRules`, the simplified, temporary calculations
that were previously happening in `importRule` are no longer necessary.

Rather than extending the existing function to support both situations,
I've taken a stab at writing a new function. It currently looks much the
same as the existing one, but I'm now going to start paring it down.