[ML] Fixes link to anomaly detection wizard from pattern analysis in Discover (#192375)
Fixes issue when clicking the `Create job using wizard` link in
Discover. Some queries cause problems with the various searches the
wizard performs.

![image](https://github.com/user-attachments/assets/79dc5920-d57d-422d-909a-61d66f3dcbe0)

This was caused by the query being added to the URL as a string when it
needs to be an object so it is correctly converted to the rison format.

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
(cherry picked from commit 4d28bd62031f57785a89103c6c46550031a38303)