[ML] Job service refactor (#191724)
Follow on from dependancy cache removal
https://github.com/elastic/kibana/pull/189729

Moving and removing as much as possible out of the `JobService` class.

- Moving the job cloning functions to a separate class as they do not
require the same dependancies as `JobService`
- Removing api basic wrapper functions like `saveNewJob`, `openJob`,
`forceStartDatafeeds` etc
- Removes toast dependency. This might be controversial, but I think it
is unnecessary for these basic job/datafeed loading functions to show a
toast if they fail. If we cannot load the basic list of jobs then there
is something else very wrong with the system. We also throw the errors,
so they will not be lost. It should be up to the calling function to
display a toast. Removing this dependency cleans up the code quite a
lot.
- The `JobCreator` classes no longer use `JobService`