[Dashboard] Fix breadcrumb update icon does not working when clicked (#192240)
## Summary

Fixes #186132 

In the `header_breadcrumbs.test.tsx`, forces the last breadcrumb
inactivity. So `onClick` did not take effect finally.

So, I'm considering putting `onClick` on the update icon.

PS: This is my first time to contribute code to kibana. If there is
anything that needs to be modified, please give me some advice and I
will try my best.

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
Co-authored-by: Hannah Mudge <Heenawter@users.noreply.github.com>