[docs] Plugin optimization, case studies (#192285)
## Summary

As titled. This PR outlines several use cases around plugin optimization
that can be helpful to engineers. This PR is part of a series, but
divided to make review easier.

It also reworks the navigation a bit to make the optimization section a
bit more clear.

(cherry picked from commit 37251b0925dd7737c55bcbde10282edd1fc96413)