WIP: Adding secondary code paths for new 'withSource' implementations
This moves the switching on feature flag to the highest level possible:
the route. Based on that determination, we call either the new
`detectionRulesClient.importRules` method, or the old `importRules`
function.

There are some type errors related to some types still being shared
between the implementations, and I need to verify that the new tests are
in the right place, but as of this commit things are plumbed through and
are mostly in the right place.