Fixes Failing test: Jest Integration Tests.x-pack/plugins/task_manager/server/integration_tests - capacity based claiming should claim tasks to full capacity (#192519)
Resolves https://github.com/elastic/kibana/issues/191117

This test became flaky again after merging this PR
https://github.com/elastic/kibana/pull/192261 which changes Kibana to
claim tasks against all partitions when the partitions have not yet been
assigned. This was due to setting the `runAt` value to `now-1000` which
means that all injected tasks were immediately overdue and claimable as
soon as possible, which means they might get claimed in different claim
cycles. This PR changes the `runAt` value to `now + 5000` so injected
tasks won't be claimed until they're all injected so they should be
claimed in one cycle.

Copied this test 60 times and was able to run the integration tests
twice (so 120 times total) with no failures.
*
https://buildkite.com/elastic/kibana-pull-request/builds/233196#0191dd0b-fe03-40cf-9fab-b211dd662993
*
https://buildkite.com/elastic/kibana-pull-request/builds/233236#0191dd94-8010-4b3c-988d-6e7d5655f989