[Alerting] Fix error when saving a rule after toggling alerts filter properties on and off (#192522)
## Summary

Closes #184170

Fixes a bug where the `alertsFilter` property gets added to `action`s
configured in the rule form when toggled on, but then never gets deleted
when toggled off. This would throw a validation error on the server for
an empty `alertsFilter` object, which is only meant to throw for
user-configured API requests and not for form usage.

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

## Release notes

Fixed a bug where toggling on and off the "If alert matches a query" or
"If alert is generated during timeframe" toggles makes it unable to save
the rule due to validation errors.

---------

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
(cherry picked from commit f22067fbc907c0ca0c5e919a25cc1afdab90d2c6)