[Synthetics] Unbreak uptime journey (#192641)
## Summary

Previously the uptime e2e journey we rely on to test the uptime overview
page was looking back 5 years to query for canned data we load during
the standard Kibana test script.

Unfortunately, that data was generated now > 5y ago, so the test started
to fail.

This patch hardcodes the date of the filter to ensure we permanently
look back to a timeframe that includes the canned data we use for
testing the page.

---------

Co-authored-by: shahzad31 <shahzad31comp@gmail.com>
(cherry picked from commit 10c1c2c7acac40edb592ed19a0c36e7ce515cf0d)