[Search] Implement RBAC Kibana feature for Search (#192130)
## Summary

This implements RBAC properly for Search. We no longer rely on the
enterprise search node to limit access to Search, so except for App
Search and Workplace Search. That means we need proper Kibana features
to manage RBAC, which this adds: one for Search and a separate one for
Relevance as that feature is guarded by an Enterprise license.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
(cherry picked from commit f7c708f135a7e6e596c3f82a671c6a048073d36e)

# Conflicts:
#	x-pack/plugins/enterprise_search/server/plugin.ts