Remove third-party testing internals from test snapshots
This removes some indavertent testing of third-party
APIs/implementations by not directly passing the result of a `render()`
call to `toMatchSnapshot()`. In all cases, this was solved by
passing `render().container` (the root element rendered by the
testing library) instead.

These changes remove implementation details of both styled-components
(#192368 being the reason for this commit) and testing-library from the
snapshots.