Define our new "validated" RuleToImport
Because we can't make the existing RuleToImport more strict (backwards
compatibility), but we need `version` to be required for prebuilt rule
customization, we now perform this check at runtime and continue to use
RuleToImport for parsing/validation.

Once the prebuilt rule epic is complete, we can replace RuleToImport
with ValidatedRuleToImport, but for now, all our new code dealing with
this new code path works with ValidatedRuleToImport.