app_search: support engine name in solution nav
Updated how we build the app search side items for the solution nav to
remove the renderItem handler and place all the engine links under the
engine name. This will only be used for the Search solution nav so we
are accepting the trade-off of losing the engine badges in the side nav
when using the solution nav.