Define DiffableRuleInput as type used for calculating rule_source
For the purposes of calculating rule_source (and mainly for diffing to
calculate `is_customized`), we now have multiple rule representations
being compared, where before all comparisons were on the much broader
`RuleResponse`, and different rule representations were converted to
`RuleResponse` before being compared.

In reality, a good number of RuleResponse fields are not used for
comparison, and as is evidenced by this commit, what's needed is (at
most) `RuleToImport` + `version` (represented now by
`ValidatedRuleToImport`). It's confusing, and unnecessary, to convert
this type to a `RuleResponse` for the purposes of comparison.

Luckily, a type system can do much of this work for us. Ideally, the
comparison code would accept the minimal representation that it needs,
and our extraction/comparison utilities would work on that. For now,
though, in order to minimize the number of changes needed, I've instead
defined the type, `DiffableRuleInput`, to be equivalent to the more
minimal input, `ValidatedRuleToImport`. Since `RuleResponse` is a
superset of that type, no existing code changes are needed.
If/when we need, we can define exactly the fields needed for comparison
as a standalone type, but for now this is a step in the right direction.