Serverless release pipeline - adjust version key for QA (#192968)
## Summary

This PR adjusts the serverless release pipeline script to use the new
key for QA when reading the version.yaml file.

Note that this is a quick fix to get the pipeline running again. Longer
term, we could have multiple slices of QA, so we should think about how
to prepare the script for that scenario.