Fix flaky test around search cancellation (#193008)
Resolves https://github.com/elastic/kibana/issues/192914

In this PR, I'm fixing the flakiness in tests where sometimes rules fail
with a different message on timeout. This is expected as it's a race
condition between the Elasticsearch request timing out and the alerting
rule getting cancelled. So we can expect one of two messages.

Note: Test is not skipped as of PR creation