[Response Ops] Fix maintenance window custom schedule create and update error (#192649)
## Summary
Fixes a bug where the backend would throw an error if we tried to create
or update a maintenance window with a custom schedule. This was due to
the `form-lib` converting everything `frequency`, `interval`, and
`customFrequency` field to a string and our logic assumed it was a
number so the `===` comparisons were failing.

### How to test:
1. Navigate to the create maintenance window form
2. Attempt to create a maintenance window with a custom schedule
3. Assert the maintenance window was created successfully
4. Attempt to edit the maintenance window with a different custom
schedule
5. Assert the maintenance window was edited successfully

Fixes: https://github.com/elastic/kibana/issues/192601

---------

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
(cherry picked from commit 134b81572c4234e9c813aa5ed5dda286a99ffc32)