Attempt to fix (or at least change) the cypress failure
The output of this failed expectation is a giant red JSON and a giant
green JSON. After manually inspecting them, they look identical except
for key order. The hope is that this change will either fix the issue or
provide a better failure, but I'm unable to even run cypress tests
locally right now.