[EDR Workflows][Osquery] Fix last results pack value (#192678)
The “Last Results” tab in the active pack details is displaying
incorrect date-time values. This is due to a mismatch in data format.
The component responsible for formatting the date-time values expects a
string, but it is currently receiving an array with a single string
inside.

Before:

![Screenshot 2024-09-12 at 10 33
39](https://github.com/user-attachments/assets/844fd699-a086-44d8-aa0b-a06e65e5aa60)

After:

![Screenshot 2024-09-12 at 12 40
04](https://github.com/user-attachments/assets/bcce9b69-58b8-438d-89e8-c78a3922845c)

(cherry picked from commit cfcb6e8f4ddf078ccdcec2a03788e39c6e4dfdef)