[Synthetics] Improve loading state for metric items (#192930)
## Summary

There are some cases where the latest addition to the overview grid
causes metric items not to load for very small numbers of monitors. This
PR aims to fix that issue, and to add a more robust loading state to the
individual items to make it easier to track when there is a request
in-flight.

Additionally, it adds a type guard to the server code and an io-ts codec
for verifying the safety of the code we use to generate our server
responses.

## Testing

Ensure that you see metric items load their duration graphs for small
numbers (3 or less) of monitors, and that scrolling etc., continues
working for larger ones.

---------

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
(cherry picked from commit 7a7888bfb8e5472d7c0656283bc1acff89ad9f5a)