[Synthetics] Increase unit test coverage (#193201)
## Summary

I frequently find when I go to modify code in Synthetics there is a lack
of good module-level coverage, which makes it challenging to confidently
change things directly, or tangential to untested code. This commonly
results in unintended regressions, or very high levels of manual
investigation during code review.

One way to get past this is to first introduce tests before modifying
the code in question, but that requires a lot of burden on the developer
trying to make the change.

We should instead find time to add additional module-level testing where
possible, and that is the spirit of this PR.

(cherry picked from commit a178d52afe99d2693bfd4c10eda8b5a9cb3ac30c)