Limit scope of exceptions test
The hope here is that this fixes some flakiness in the test, by
restricting the alerts being asserted upon to only those created within
the test (via filtering on the created rule's ID). Surrounding tests do
this, and it seems likely that the failure
(https://github.com/elastic/kibana/issues/181887) was caused by a
previous test's alerts being present during this test.