WIP: Separating route-specific logic from general importing
This is the first pass of keeping any route-specific logic out of the
DetectionRulesClient, and particularly the new `importRules` method
that's added in this effort.

It removes route concepts like streams and status codes and responses
(except RuleResponse) from the client method, and moves them into an
intermediate function that then calls the client.

I'm also leveraging the new `RuleImportError` here, and `importRules`
now returns  `Array<RuleResponse | RuleImportError>`, which feels a lot
better.

At this point, tests are very much broken, so once this idea is vetted
I'll be fixing those.