[ResponseOps][Cases] Fix a bug with cases telemetry where data from other spaces are not included (#193166)
## Summary

The Find SO API supports the `namespaces` parameter where you can define
the spaces that the SO client should search for. If you omit the
`namespaces` parameter, the SO client will use the active space. This PR
creates a wrapper around the SO client to add the `namespaces: ['*']` to
all Find SO usages to count telemetry on all spaces.

### Checklist

Delete any items that are not applicable to this PR.

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

### For maintainers

- [x] This was checked for breaking API changes and was [labeled
appropriately](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>