Clean up some interfaces by combining two helper utilities
As evidenced by the unit tests, we were never calling
`createRulesStreamFromNdJson` without also calling
`createPromiseFromStreams` to get a workable promise.

As a consequence of calling these separately, callers had to manually
cast the result of `createPromiseFromStreams`, which made for a lot of
unnecessary verbosity.

We now instead simply call `createPromiseFromRuleImportStream`, and
downstream functions can now deal with the simpler shared type
`RuleFromImportStream` directly.