PrebuiltRulesImportHelper -> RuleSourceImporter
This is a significant refactor of the responsibilities of The Class
Formerly Known as PrebuiltRulesImportHelper.

Besides the (hopefully) more descriptive name (as the class is
responsible with generating appropriate `rule_source` values during
import), the class now combines what were previously multiple sequential
method calls into a single one (still `setup()`), while adding some
additional methods to aid in `rule_source` calculation.

The big one (and the one that incited this change) was the ability to
ask of an incoming (importing) rule: does this represent a prebuilt
rule, or not? This is self-evident from a calculated `rule_source` (i.e.
`rule_source.type === 'internal'`), but for this use case (defaulting a
prebuilt rule's version) we need to do so _before_ calculating the full
`rule_source`.

In order to make this determination, we need our list of rule assets,
which were already fetched by the instance. But rather than require the
caller to take those assets and pass them to another helper function, I
opted instead to move that logic to a method on the RuleSourceImporter.

Similarly, I replaced the manual invocation of
`calculateRuleSourceForImport` with a (wrapping) method on the
RuleSourceImporter.